feat: refactor social domain to use Long IDs and add query services

+++ src/main/java/com/nexora/nexorabackend/social/domain/commands/CreateLikeCommand.java
@@ -2,7 +2,7 @@ package com.nexora.nexorabackend.social.domain.commands;
 
 import java.util.Objects;
 
-public record CreateLikeCommand (Integer userId, Integer postId) {
+public record CreateLikeCommand (Long userId, Long postId) {
 
     public CreateLikeCommand {
         Objects.requireNonNull(userId, "User ID cannot be null");
