+++ src/main/java/com/nexora/nexorabackend/social/interfaces/rest/resources/CreateLikeResource.java
@@ -1,8 +1,8 @@
 package com.nexora.nexorabackend.social.interfaces.rest.resources;
 
 public record CreateLikeResource (
-    Integer userId,
-    Integer postId 
+    Long userId,
+    Long postId
 ){
     
 }
