+++ src/main/java/com/nexora/nexorabackend/social/interfaces/rest/resources/CreatePostResource.java
@@ -3,10 +3,10 @@ package com.nexora.nexorabackend.social.interfaces.rest.resources;
 
 public record CreatePostResource (
     String title,
-    Integer authorId,
+    Long authorId,
     String body,
     Integer reactions,
     Integer categoryId,
     Integer fileId,
-    Integer communityId 
+    Integer communityId
 ){}
