+++ src/main/java/com/nexora/nexorabackend/social/interfaces/rest/resources/PostResource.java
@@ -4,7 +4,7 @@ package com.nexora.nexorabackend.social.interfaces.rest.resources;
 public record PostResource (
     Long id,
     String title,
-    Integer authorId,
+    Long authorId,
     String body,
     Integer reactions,
     Integer categoryId,
